refactor: keys: Use Dict for KEY_BINDINGS and HELP_CATEGORIES.
KEY_BINDINGS was initialized as an OrderedDict to preserve its
insertion order when used in lint-hotkeys.read_help_categories
to generate hotkeys.md and views to render a HelpView.

HELP_CATEGORIES was initialized as an OrderedDict to preserve its
insertion order when used in lint-hotkeys.get_hotkeys_file_string
to generate hotkeys.md and views to render a HelpView.

Since Dict also preserves insertion order from python3.7, we can
safely initialize both as a Dict without any change in behavior.