model/views: Fetch next unread topic using current message state.
This commit changes the behavior of the get_next_unread_topic method to
use the current message state (ie. the current topic) in calculating the
next unread topic to be cycled to.

If there is no current message available, ie. an empty narrow with no
focus, the logic attempts to use the current narrow to determine the
best course of action.

This replaces the previous approach of using a _last_unread_topic stored
in the model.

Tests updated.