lint-all: Reorder linting for faster development cycle.
By running more important, faster checks first, feedback is given to the
developer sooner when failures occur.

In addition, this addresses two specific issues:
- ruff and black appear to disagree in some cases; running them first makes
  that faster to resolve
- isort gives less specific feedback on error locations than black (or ruff)
  when there are syntax errors, so if isort is first and fails, causing
  linting as a whole to fail, then that information is not available

The tools most likely to trigger failures are now run in order of the
fastest tool first, ie. starting with:
- ruff (check) [was 3rd]
- black (check) [was 4th]
- isort (check) [was 1st]
- mypy [was 2nd]

Spelling and code-document synchronization typically run fast, but are less
important checks that can be worked on later, so remain after the above
tools.