ui/buttons: On exiting topics view, associate topic name with stream_id.
This is achieved through a new method associate_stream_with_topic in the
View, which saves to a new internal dict.

This will support later restoring the topic position in the UI.

Since the topic list may be reordered between saving and restoring the
state, the current index in the topic list is insufficient; the name of
the topic is used intead.