views: views: Add a function _focus_position_for_topic_name in TopicsView.
This function returns the index of the topic list that needs to be
focussed.
Instead of writing it directly in the init function where it
is called,it is written as a separate function for easy/clear
testing.
The function loops over self.log but when the desired topic name
is found it returns the index.
The index is then used to update the value of focus_position of listbox.
Instead of using set_focus function,focus_postion is used for easy testing.