views: Update user info popup to display custom profile fields.
This commit updates the user information popup to display the custom
profile fields stored in the model, including processing additional data
received from the model in _fetch_user_data.

All fields except the Person picker field are strings, so if a Person
picker field is present, it is converted to a string of usernames
separate by commas.

Tests added and updated.

Fixes #1338.