helper: Match topic search text against start of all words in topic.
This extends the existing behavior which only matched against the start of
the topic, and applies in entering the topic in the compose box as well as
linking to streams and topics in message content.

A similar approach is taken as with streams, splitting on spaces as well as
a set of additional delimiters (-_/).

The topics fixture is extended to include text split by delimiters.
Test cases updated and extended.

Fixes #1358.