model/api_types/helper: Add _clean_and_order_custom_profile_data method.
This commit adds the _clean_and_order_custom_profile_data function to
model.py. The function takes the profile data (in Zulip API format) as
input and processes fields to get label, value, type and order of the
field. It also sorts profile data according to the "order" attribute.

The value of field type 6 (Person picker) is stored as a list of user
ids. The values for all other field types are stored as strings.

Data structures related to custom profile fields are added to
api_types.py (CustomProfileField, CustomFieldValue) and helper.py
(CustomProfileData).

Test and related fixtures added.