test_ui_tools: Tests added for the _focus_position_for_topic_name.
Here the saved_topic_state parameter gives the value of the focused that
should be focused(if haven't been selected than None).
In the function body a list_box is created with its body's topic_names are
assigned. Then log is created using urwid library. Then the
saved_topic_in_stream_id is set by patching the parameter.
After that the actual function takes place and the value of
focus_position is assigned and asserted.