boxes/messages: Disable recpient editing when editing a private message.
In order to fully disable editing (not only cycling to it, but e.g.
prevent editing through mouse clicks) we need to make the editing
of a message a different UI box than the usual reply.

This new box uses a urwid.Text to display the recipient information
rather than providing an editable field.

Fixes #774.

Co-authored-by: zormit <nt4u@kpvn.de>