boxes: Disable cycling while editing a Private Message.
In other words: CYCLE_COMPOSE_FOCUS is now deactivated for the case of
editing a private message.

This is one core change of this PR and solves the issue that it doesn't
particularly make sense to move the focus to the recipients. Note that
we don't yet disallow *clicking* into the recpient list, which will be
a later change.

Co-authored-by: zormit <nt4u@kpvn.de>