README: Rewrite gitlint section to enhance use of the tool.
Prior to this commit, while the tool is installed in every development
environment, installing the hook was simply 'highly suggested'.

This rephrases the text to improve the motivation for using gitlint,
including running it manually, and then showing how automating it is
therefore useful.

This section still follows the general commit message style section; it
may be necessary to refer one to another, or otherwise combine them to
improve motivation towards using this tool. Alternatively, gitlint could
be added to CI.