refactor: api_types/model: Improve typing of Subscription change types.
This splits the previous SubscriptionEvent into
- SubscriptionUpdateEvent (op:update): personal settings
- SubscriptionPeerAddRemoveEvent (op:peer_add/remove): subscribers

SubscriptionSettingChange is defined for reuse both with the
SubscriptionUpdateEvent and for calls to update_subscription_settings,
which is applied in the model.

Specific (un)supported personal subscription settings are detailed and
enforced through comments and a new PersonalSubscriptionSetting Literal.

A spurious message_ids field is removed, likely from when the Event
object described all possible fields.