refactor: api_types/model: Split & improve update_message event type(s).
All fields specific to update_message events are documented, even if
many are commented due to being unused at this time.

This event is quite complex, with sets of fields only present when
certain updates are made. To simplify the expression of this condition,
the original event is split into a base type (fields always present) and
two derived types (UpdateMessageContentEvent,
UpdateMessagesLocationEvent).

Since a given update_message event can be one or both of these types,
the _handle_update_message_event method uses casting based on the
presence of fields unique to each type of event.