bugfix: api_types/model: Correctly rerender when /me status changes.
Previously the is_me_message flag in messages were not updated locally,
if the content was updated to be recognized as such by the server.
Such a flag is provided directly in update_message events
(is_me_message) as in message objects fetched from the server, but was
not synchronized until now.

This had no effect on /me messages which were edited to lose that
status, since the message rendering code did not find the matching /me
message content and so updated the rendering to show as a regular
message.

However, for messages which were initially not /me messages, the
/me rendering path could never be entered if they were subsequently
edited to become /me messages, since the property of the message was not
updated to allow this.

Tests updated and new test cases added.