run/keys/core/boxes: Allow editing message content via external editors.
A new OPEN_EXTERNAL_EDITOR command (hotkey: Ctrl o) may now open an
external editor, if details of one are provided, to allow message
content to be edited outside of the application, during message compose
or editing.

The external editor may be specified by, in order of increasing
precedence:
- the semi-standard EDITOR environment variable
- a ZULIP_EDIT_COMMAND environment variable
- a new 'editor' key in the zuliprc file

ZULIP_EDIT_COMMAND is added to allow a custom form of EDITOR to be used,
ie. to avoid causing issues with other tools that may depend upon the
latter.

The command is validated as being set and the path to it as existing,
with errors reported to the user.

The feature then operates via a randomly-named temporary file,
initialized with any existing message content, which is opened by the
application with the editor. The application pauses at this point, until
the editing is complete, when the content is substituted back into the
compose content area, and the application resumes.

README updated to document new zuliprc option.
hotkeys automaically updated for new command.

Tests updated.

Additional defensive approaches and textual changes made by neiljp.