buttons: Ensure foreground on stream background is never transparent.
Previously the foreground color on a stream-color background was the
general theme background color, or "black" if that general background
color was Falsey - typically the empty string: "".

The empty string is one value to indicate the terminal default color in
urwid - the background in this case - and so enable a transparent
background in some terminal emulators. This logic therefore ensured that
the text color on a colored background was not itself transparent, and
enabled the previous implementation of background transparency.

However, it is clearer if the alternative explicit string of "default"
is used for this purpose, since this is a finite recognizable string
that can be structured in the same way as other multi-color-depth colors
are now specified - as a sequence of color words. This has not been an
issue ever since transparency support was removed and the theme system
dramatically updated, but is necessary to add back support for
background transparency.

This commit therefore extends the transparent detection for such cases
to include "default", without which use of "default" elsewhere would
lead to transparent foreground characters.