lister/run-mypy/pyproject: Improve follow-imports & type-check lister.
lister.py was previously explicitly excluded from type-checking, but if
we remove the hard-coded `follow-imports=silent` in tools/run-mypy, it
is evident that it was previously being checked due to being imported
where it was used - but error messages were not reported.

With minimal typing and a minor change to lister.py, the file can be
made to pass type-checking, enabling an overall move to the better
default `follow-imports=normal` setting in pyproject.toml. Moreover,
this also allows lister.py to be removed from the list of files excluded
from type-checking we maintain in pyproject.toml and run-mypy.