pyproject/tools: Error on missing types in libraries, with exceptions.
Rather than mypy simply ignoring any missing type information on
imports, this change sets this to be an error by default.

Separate sections for imports used in the main application and only
during development are added, to exclude certain libraries (import
patterns) from resulting in errors, where the type information is not
available. This makes the lack of typing more explicit and limited.

In addition, an import in tools/convert-unicode-emoji-data is marked
with a type-ignore comment, since the imported file is only temporary.