run-mypy: Add support for repo/project-specific mypy arguments.
In general this behavior is now handled in pyproject.toml instead, but
in certain cases this is not possible.

This approach enables options to be applied differently when running
mypy on repos/projects (folders) which interact, eg. import from each
other, and changing one folder of files would otherwise be necessary
*only* to satisfy mypy when running on the other.