color/themes/zt_dark: Enable marking common background color in styles.
This allows easily updating the background color to a theme by updating
the color in only one location, as well as making it clear upon reading,
if a background to a style is distinct from the background.

To apply this to a theme with some common background color `Color.BG`:
- import `Background` from `zulipterminal.config.color`
- replace `Color.BG` values with `Background.COLOR` in styles (STYLES)
- add a "background" entry to the META dictionary, with value `Color.BG`

This commit applies this to only the 'zt_dark' theme.

Note that in the absence of a globally reasonable fallback value, if the
"background" entry is not specified in META, the background will be
transparent or the terminal default background color.

Tests updated.