run/README: Configuration infrastructure to enable/disable transparency.
This hooks up the underlying feature from the previous commit to
new configuration options possible in the zuliprc file, which can be
overridden by command-line options.

These take the same form as for notifications and autohide, with
- `transparency=enabled|disabled` in zuliprc
- `--transparency` or `-no-transparency` on the command-line

The default value for this option is to be disabled.

This value is not currently passed to the Controller, since the theme is
generated in run.py.

README updated to indicate new transparency configuration option.

Tests updated.