model: Improve typing of stream_dict.
With the revamp of the Subscription TypedDict in an earlier commit, we
use it to improve the typing of stream_dict in this commit.

The `is_old_stream` is removed from stream fixtures since ZFL 14
removes them from the Streams object. This field was always
equivalent to stream_weekly_traffic != null on the same object.

Tests and fixtures updated to support the new typing.