views: Introduce FileUploadView to upload files.
FileUploadView's base class is PopUpView. On keypress of `FILE_UPLOAD`
it passes the value of the edit_text to _handle_file_upload function.
The __init__ have write_box as a parameter which is used to insert the
URI in the write_box.The PopUpView is closed on 'ESC' and the upload is
done via 'ENTER'.