api_types: Update Stream and Subscription typeddict fields.
This commit updates the `date_created` and
`stream_post_policy` fields present in the Stream typeddict,
and the `role` field in the Subscription typeddict.

For `date_created`, since servers with ZFL<30 do not include
the field, but ZT does add it with a value of None to ensure
consistency, the type is changed to
`NotRequired[Optional[int]]`.

For `stream_post_policy`, the field is only included after
Zulip 3.0 / ZFL 1, so it has been changed to a NotRequired
field.

For `role`, the field has been removed in Zulip 6.0 (ZFL 133),
so it has been removed from the Subscriptions typeddict and
also from the unsubscribed_streams_fixture.