api_types: Update Stream and Subscription typeddict fields.
This commit updates the `date_created` and `role` fields present in the
Stream and Subscription typeddicts respectively.

For `date_created`, since servers with ZFL<30 do not include the field,
but ZT does add it with a value of None to ensure consistency, the type
is changed to `NotRequired[Optional[int]]`.

For `role`, the field has been removed in Zulip 6.0 (ZFL 133), so it has
been removed from the Subscriptions typeddict.