symbols: Add symbols (markers) for use with common narrows.
The 'Mentioned messages' and 'Starred messages' symbols are set to ASCII
characters, similar to those in the Zulip web app.

The selected 'All messages' symbol is the closest match found after some
testing, and appears to be available fairly widely. An appropriate
descriptive comment is added, as with other non-ASCII symbols, for later
reference.