test_popups: Tests for FileUploadView.
Firstly the TestWriteBox is imported as unlike other popups
FileUploadView takes a argument write_box. A fixture write_box
is created which creates a instance of TestWriteBox and
returns its write_box fixture.
There are two cases of keypress test- one for keys which dosent make
the popup exit and the other for key that exits the popup.
There are 4 testcases for _handle_file_upload. The tests are based on
if the return of get_file_upload_uri is None or a valid uri. On that
basis the errors/append_func are called or not is asserted.