buttons/ui_sizes: Add prefix symbols to narrowing buttons in top-left.
This commit adds prefix symbols to the main buttons in the top left
corner of the UI. This aims to make the main buttons stand out more and
approximate the designs used in the web app.

This change also leads to a cleaner design, since these top buttons are
now indented and aligned similarly to the panels beneath them, eg. the
list of streams.

The symbol used as a prefix in headers of direct messages is reused for
the 'Direct messages' button, with other buttons using symbols added in
the previous commit.

The left part of the UI is increased in width to accommodate the new
additions.

Note that the "title" style is used to make the icons bolder, though
this should be decoupled in future.