refactor: index active_conversation_info by id.
The active_converstation_information is a Dict where id of the users is
the key and the full_name of the user is the value.
Previously the indexing was done by name which is changed
to id now because of the possibility of non-unique names
present in the active_conversation_info.
In core.py active_conversation_info gets the full_names of the emails
 present in self.active_conversation_info's values.