messages: Show prefix symbols before the title of the current narrow.
As previously done to other parts of the UI, this commit adds prefix symbols
to the narrow title (before the message search input area).

This is applied to all supported narrows:
- all messages
- all direct messages
- mentions
- starred messages
- single stream
- single topic
- direct message conversations (1:1 and group)

The spacing around the prefixes is laid out to be similar to that for
message headings, so that the icons line up cleanly.

This is intended to make the UI more similar to the web app UI and unify
the overall look.

Tests adjusted accordingly.