lint-and-test: Revert previous change to use latest python on MacOS.
Using '3.x' automatically selected a high single version of python on
MacOS.

However, it bumped the python version over the version that we support,
breaking with no warning - and also using different python versions
between PRs.

This could be useful to have with a manual trigger, or if it did not
break regular builds.

This is an effective revert of d8bae8894544e4cdaafdd1b5aea6682e42b0f854.