core/boxes: Improve handling of pressing Esc during message compose.
Introduces variable MAX_MESSAGE_LENGTH_CONFIRMATION_POPUP to set
a threshold for maximum length of message in compose box beyond
which it prompts a confirmation popup instead of the current
instant exit when Esc is pressed.

Call _set_default_footer_after_autocomplete() in the exit_compose_box
method to reset the footer in case autocomplete was in use.

Fixes #1342.