boxes/messages: Extend narrow title colored background area.
The colored background area of stream message headings are already
visually balanced with a leading and trailing space. The narrow title
prefixes added in the previous commit contain a leading space to align
the prefixes with those in message headings; this commit adds a trailing
space, which balances the colored area in a similar way to the stream
message headings and avoids an abrupt end to the narrow text.

The existing spacing between the narrow title and message search text is
maintained by removing one of the two spaces between the `Search [/]:`
and the narrow title with a general background color, and substituting
it with an additional space at the end of each narrow title.

Tests adjusted accordingly.