bugfix: views: Stream email is not in subscription objects from ZFL226.
This change avoids a crash related to the removal of the email_address
field from subscription objects at feature level 226 (ZFL226) onwards.

The minimal fix here is to indicate the email is unavailable, if this
situation is detected.

The change at ZFL226 (Zulip 7.5 and 8.x) enables fetching of the email
address of a stream on demand, which we can use later to dynamically
fetch the data and insert it into the popup. Since that call could fail,
this error handling is likely to remain useful.

Test generalized to include previous and new behavior.