lint-and-test: Check commit style for commits on PR branches.
This requires an approach similar to the isolated-commits job, since the
checkout action by default only fetches the latest state of the code.

This job is also similar in that since it applies to a branch, it is
limited to pull requests.

This uses the new gitlint install option, to minimize necessary package
installs using the current method.