lint-and-test: Use testing_minimal instead of full pytest install.
In one case we were installing the `testing` option fully, which we
shouldn't need for linting.

In another we were manually installing pytest using pip at its latest
version, which could vary subtly from the version we use otherwise.

This change ensures the version of pytest we use is consistent and
somewhat more minimal when installed in CI.