requirements[dev]: Upgrade mypy from ~=1.5.0 to ~=1.6.0.
This requires an update of a type ignore in convert-unicode-emoji-data,
which is necessary due to the imported file being conditionally present.

CI and local Python 3.9 require different type ignore specifiers, so
this also adds 'unused-ignore', to avoid errors when the other type
ignore is apparently not used on the other platform.