bugfix: messages: Fix crash when editing message.
When `realm_message_content_edit_limit_seconds` is None, which now means
no time limit for editing messages, ZT crashes with a `TypeError`.

This was diagnosed in Zulip cloud, but may be triggered for any version
of the server from feature level 138, which was released first in Zulip
server 6.0. Prior to this feature level, no limit on editing messages
was represented by a numerical value of zero; a direct numerical
comparison was therefore possible, but now causes an error.

Fixes #1467.