bugfix: model: Set pointer to message index instead of message id.
Currently, the narrow pointer is set to the anchor message id when
fetching messages from the server in the get_messages method. This
contradicts with its use in functions - get_focus_in_current_narrow and
set_focus_in_current_narrow, which treat it as an index pointing to the
last focused message in the narrow's message list. This commit makes the
pointer consistent with its current use as a message index instead of a
message id by setting it to the anchor message index in the message
list.

Fixes #1226.