tests: model: Fix test for get_messages method with false first_anchor.
This commit fixes the test_get_message_false_first_anchor test.
When the first_anchor parameter is set to False in the get_messages
function, we do not assign any value to the current narrow's message
pointer. This is represented by the change in the assert statement in
the test.

Also, setting the messages_successful_response's anchor to 0 before
initializing the Model in the test leads to the pointer being set to 0,
which is not intended and is moved to after initializing the Model.