bugfix: core: Resolve macOS crash on opening message in web browser.
This occurs due to an attribute error in the existing code, since a
class in the standard library webbrowser module didn't have a 'name'
attribute until Python 3.11 under macOS.

While this attribute was not officially documented until Python 3.11,
many if not all other browser drivers provided a name attribute before
this point, and the existing use of the attribute did not cause issues
when our original implementation was tested.

This bugfix substitutes a default value in the absence of the name,
rather than testing against the specific platform/python combination,
which will protect against similar issues arising in future.

Test updated.

Fixes #1471.