hotkeys: Refined UI display of hotkeys.
Update all the files that display hotkeys to the user,
to use the new function and representation.

The internal usage (mainly for keypresses)
is left to use the default Urwid representation.

Hotkeys are displayed in the buttons, the help menu and the footer.
Updated them and their test mocks.

The spacing has been adjusted to accomodate
for the length of the keyboard_key_for_urwid_key() and
primary_keyboard_key_for_command functions.