views: Enhance the behaviour of `Esc` in emoji picker popup.
Previously, pressing Esc anytime when the emoji picker
is in use will exit the popup.

Now, different behaviours are supported depending on the state.

Escaping from a non-empty searchbox clears it.
Escaping from an empty searchbox closes the popup.

Escaping from the emoji list exits you out of the popup,
if the selected emoji button was toggled after the last search.
Otherwise, it takes you back to the searchbox
and retains the last search text.