views: Make the ALL_MESSAGES command work globally.
This command worked only when a message was selected, using it as
an anchor to fetch messages.
Now, it has been made consistent with the other narrow commands,
to work from any panel.

This could not be implemented in ui.py along with other narrow commands,
because of the conflict caused by the Esc key also being assigned to
reset search in the side panels (GO_BACK command).

When both operations,
1. reset search in the current panel view
2. narrow to all messages
are possible, pressing `Esc` is set to trigger only
the reset search operation and not the ALL_MESSAGES command.
The next keypress of `Esc` will go to the home view once the
current panel view is restored to its default state.

Fixes #1505.