refactor: run: Set zuliprc_path through fetch_zuliprc().
Previously, zuliprc_path was set directly in main().
Now, we are setting it 2 levels deeper, inside of fetch_zuliprc(),
passing it to main() through parse_zuliprc().

main() now sets the default value of zuliprc_path to an empty string,
and uses the returned path from parse_zuliprc().

This is done to allow adding support for multiple zuliprc paths later.