run: Add function to get zuliprc path, given realm name prefix.
To fetch the zuliprc from a realm's directories in config,
we will add functionality to search for matching realm names
given their prefix.

If there are several realms with the same prefix,
or none with a matching prefix,
or the matching realm does not contain a zuliprc,
then exit with an error message.