run: Use the new config directory as a default zuliprc path.
In addition to the previous default path ~/zuliprc.

Introduce the .config path of zulip-terminal as a str constant.

Add a function check_for_default_zuliprc() to search for zuliprc files
in both the default locations, when the --config-file argument is not
used.

It is added to ensure that users who have only a single account can
skip using the account alias argument that will be added.

Docstring updated.
Tests updated.