refactor: run: Stop passing zuliprc_path to login_and_save().
Previously, we had to pass the zuliprc_path because the --config-file
option could be used to create a zuliprc at a custom location.

Now, all accounts are only saved to the standard config path. So, by
moving the print statement to the outer function, we can simplify
login_and_save().