requirements: Upgrade lxml from ~=4.9.2 to ==4.9.4.
The 5.x releases of lxml should support Python 3.7 in general, but it
appears that 4.9.4 is the latest release for which PyPy 3.7 builds in
CI, the last release in the 4.x branch.

Dropping support for Python 3.7 will allow progress towards the latest
lxml. This should then allow inclusion of PyPy 3.11 in CI, which
compiles in CI with at least lxml 5.3.1 (maybe earlier), but not
<=4.9.4.